Feature 'multiple MTR jobs'
This commit introduces a feature to trigger multiple
mtr-jobs. One goal is to keep the existing user interface
intact while providing the new feature.

Added flag: -label <job-label>

In order to distinguish multiple running jobs in the resulting
Prometheus /metrics response, each job needs a distinct label.

Added flag: -jobs <path-to-jobs-file>

To define multiple MTR jobs, the concep of a "jobs file" is introduced.
It contains, per line, the label for the job (equivalent for -label),
the schedule upon which the job is launched (equivalent to -schedule)
and the flags for `mtr`.

Added flag: -watch-jobs <schedule>

The -watch-jobs flag makes `mtr-exporter` to reparse the file given
via the -jobs flag at the defined schedule. When change is detected,
all previous jobs are stopped and the new jobs are put into effect.
If no changes are detected, nothing happens.

Note: No INOTIFY or other similar mechanic are used as of now to
keep cross-platform nature of `mtr-exporter` in place without
introducing new dependencies, build-changes etc.