Add -flag.deprecatedMetrics to control exposure of deprecated metrics
Via 7cd5e4f, a bunch of metric names were marked as "deprecated".
This is mainly relevant for users who rely in their alert/monitoring
setups on the (now deprecated) metric names. To make the transition
period easier, the deprecated metrics can still be exposed by using
the -flag.deprecatedMetrics.

For users neither in the need nor the mood to use the old metrics at
all, the default is to NOT show the deprecated metrics.