Container image: change container entrypoint to reap child processes
`mtr-exporter` spawns (lots) of `mtr` processes over its runtime
lifetime. Each of those processes terminates and will linger around as
"zombie" processes if not properly reaped. Usually, Pid-1 inside of
a container is taking care about this.

Up to this commit, `mtr-exporter` was Pid-1 but lacked the feature to
reap its spawned children. `docker` provides an `--init` flag which is
nice but puts the burden of remembering that detail to the operator of
the container.

This commit is improving the situation by utilizing the help of `tini`
to run as container entrypoint and thus reaping spawned children
properly. As a result: no "zombie" processes anymore.

Related: https://github.com/mgumz/mtr-exporter/issues/17