Build system: more robust container-image builds
A container image usually consists of a image name and a tag.
The tag consists usually consists only of the version number. Sometimes,
information like a "pre-release" or a "build-number" are added to the
tag. Example given:

  mtr-exporter-0.3.0-rc1

The build system tries to build a binary named

  mtr-exporter-0.3.0-rc1.linux.amd64

and fails because the rule to build such a binary does not exist.

This commit fixes the issue by relaxing the wildcard pattern.