Add mtr_runs_total metric and track failed runs
The new metric `mtr_runs_total` keeps track of how often a
certain job of `mtr` was launched / returned. To allow detecting
issues in the network (for which `mtr` returns "only" an error message),
the error message of `mtr` is used as "error" label. A successful run
is lacking the "error" label. Thus:

* `sum by(mtr_exporter_job)(mtr_runs_total{})` provides the absolute total
  number of `mtr` runs
* `sum by(mtr_exporter, error)(mtr_runs_total{error!=""})` provides the
  failed runs
* the diff between the previous queries are the amount of successful
  runs

The error message of `mtr` is "normalized": all chars in the original
`mtr` message are lower cased, only [0-9,a-z,- ] are kept, everything
else is mapped to `-`.

Also, when no job has returned something yet, the metric exporter
would not report uninitialized values.