Improve handling zero hubs - add mtr_report_min_loss
In some situations, `mtr` probes a path and gains no response
within --first-ttl and --max-ttl … for whatever reasons (network
operators sometimes have interesting oppinions about ICMP).

It is a not a bug of `mtr` to exit with an error in this case: it
sends out a packet and checks what comes back. As of today, `mtr`
does not include what it tried in the --json output. The only
way to deduce that kind of all packets remained unanswered is to
check the amount of `.hubs[]` - which is 0.

To give a user of `mtr-exporter` something to work with: this commit
introduces a new metric: `mtr_report_min_loss`. It reports to minimal
percentage of packet loss observed. If it's value is 100: all packets
were lost. If not … then not :)