Feature: add label `path_id`
From time to time, the path ("route") to a given destination changes.
This can have direct impact to the observed latency. The regular
visualisation might be just based upon the metric `mtr_report_loss`.

One problem with that is its hard to detect a path change. While the
latency per hop is provided as an individual metric with labels `host`,
`count`, `hop=first|intermediat|last`, it is really tricky to
query Prometheus for the path itself. Keep in mind: the most
likely hosts changing along the route towards the destination is
mostly somewhere in the middle (the "inter" in "internet").

This commit adds a new label `path_id` onto the metrics which also have
the `hop="last"` label set. Technically it is the Blake2b with 64 bit
digest size of the concatenation of the hosts involved.

Thus, one can query an aggregate, for example

   max by (dst, path_id) (mtr_report_loss{hop="last"})